Rename docs files to specify instruction set
