+++ .eslintignore
@@ -6,5 +6,5 @@ out
 views
 
 # Autogenerated files
-lib/handlers/asm-docs.js
-lib/handlers/asm-docs-arm.js
+lib/handlers/asm-docs-amd64.js
+lib/handlers/asm-docs-arm32.js
