+++ lib/handlers/api.js
@@ -31,7 +31,7 @@ import { logger } from '../logger';
 import { getShortenerTypeByKey } from '../shortener';
 import * as utils from '../utils';
 
-import { AsmDocsHandler } from './asm-docs-api-arm';
+import { AsmDocsHandler } from './asm-docs-api-arm32';
 import { Formatter } from './formatting';
 
 export class ApiHandler {
