+++ lib/handlers/asm-docs-amd64.js
