+++ lib/handlers/asm-docs-api-amd64.js
@@ -24,7 +24,7 @@
 
 import * as props from '../properties';
 
-import { getAsmOpcode } from './asm-docs';
+import { getAsmOpcode } from './asm-docs-amd64';
 
 export class AsmDocsHandler {
     constructor() {
