+++ lib/handlers/asm-docs-api-arm32.js
@@ -24,7 +24,7 @@
 
 import * as props from '../properties';
 
-import { getAsmOpcode } from './asm-docs-arm';
+import { getAsmOpcode } from './asm-docs-arm32';
 
 export class AsmDocsHandler {
     constructor() {
